refactor dailyGarment so more is in initial container
